Split test cases into separate files
